ints and floats

+++ src/com/uwsoft/editor/renderer/data/ResolutionEntryVO.java
@@ -19,9 +19,9 @@ public class ResolutionEntryVO {
     public float getMultiplier(ResolutionEntryVO originalResolution) {
         float mul;
         if(base == 0) {
-            mul = originalResolution.width/width;
+            mul = (float)originalResolution.width/width;
         } else {
-            mul = originalResolution.height/height;
+            mul = (float)originalResolution.height/height;
         }
         return mul;
     }
